FII-030 - Display the Inline files on the File inspector - UI implementation (LEAN-920)
